+++ handlers/handlersClientes.go
@@ -11,9 +11,8 @@ import (
 )
 
 func CreateClienteHandler(w http.ResponseWriter, r *http.Request) {
-	sec.IsAuthenticated(w, r)
-	log.Println("Create Cliente")
-	if r.Method == "POST" {
+	if r.Method == "POST" && sec.IsAuthenticated(w, r) {
+		log.Println("Create Cliente")
 		name := r.FormValue("Name")
 		endereco := r.FormValue("Endereco")
 		capacidade := r.FormValue("Capacidade")
@@ -27,14 +26,15 @@ func CreateClienteHandler(w http.ResponseWriter, r *http.Request) {
 		}
 		sec.CheckInternalServerError(err, w)
 		//log.Println("INSERT: Id: " + strconv.Itoa(id) + " | Name: " + name + " | Code: " + code)
+		http.Redirect(w, r, route.ClientesRoute, 301)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	http.Redirect(w, r, route.ClientesRoute, 301)
 }
 
 func UpdateClienteHandler(w http.ResponseWriter, r *http.Request) {
-	sec.IsAuthenticated(w, r)
-	log.Println("Update Cliente")
-	if r.Method == "POST" {
+	if r.Method == "POST" && sec.IsAuthenticated(w, r) {
+		log.Println("Update Cliente")
 		id := r.FormValue("Id")
 		name := r.FormValue("Name")
 		endereco := r.FormValue("Endereco")
@@ -49,14 +49,15 @@ func UpdateClienteHandler(w http.ResponseWriter, r *http.Request) {
 		sec.CheckInternalServerError(err, w)
 		updtForm.Exec(name, endereco, capacidade, cnpj, id)
 		log.Println("UPDATE: Id: " + id + " | Name: " + name + " | cnpj: " + cnpj)
+		http.Redirect(w, r, route.ClientesRoute, 301)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	http.Redirect(w, r, route.ClientesRoute, 301)
 }
 
 func DeleteClienteHandler(w http.ResponseWriter, r *http.Request) {
-	sec.IsAuthenticated(w, r)
-	log.Println("Delete Cliente")
-	if r.Method == "POST" {
+	if r.Method == "POST" && sec.IsAuthenticated(w, r) {
+		log.Println("Delete Cliente")
 		id := r.FormValue("Id")
 		sqlStatement := "DELETE FROM clientes WHERE id=$1"
 		deleteForm, err := Db.Prepare(sqlStatement)
@@ -66,8 +67,10 @@ func DeleteClienteHandler(w http.ResponseWriter, r *http.Request) {
 		deleteForm.Exec(id)
 		sec.CheckInternalServerError(err, w)
 		log.Println("DELETE: Id: " + id)
+		http.Redirect(w, r, route.ClientesRoute, 301)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	http.Redirect(w, r, route.ClientesRoute, 301)
 }
 
 func DeleteClientesByRoleHandler(roleId string) {
@@ -81,28 +84,30 @@ func DeleteClientesByRoleHandler(roleId string) {
 }
 
 func ListClientesHandler(w http.ResponseWriter, r *http.Request) {
-	log.Println("List Clientes")
-	sec.IsAuthenticated(w, r)
-	rows, err := Db.Query("SELECT name, endereco, capacidade, cnpj FROM clientes order by id asc")
-	sec.CheckInternalServerError(err, w)
-	var clientes []mdl.Cliente
-	var cliente mdl.Cliente
-	var i = 1
-	for rows.Next() {
-		err = rows.Scan(&cliente.Name, &cliente.Endereco, &cliente.Capacidade, &cliente.Cnpj)
+	if sec.IsAuthenticated(w, r) {
+		log.Println("List Clientes")
+		rows, err := Db.Query("SELECT name, endereco, capacidade, cnpj FROM clientes order by id asc")
 		sec.CheckInternalServerError(err, w)
-		cliente.Order = i
-		i++
-		clientes = append(clientes, cliente)
+		var clientes []mdl.Cliente
+		var cliente mdl.Cliente
+		var i = 1
+		for rows.Next() {
+			err = rows.Scan(&cliente.Name, &cliente.Endereco, &cliente.Capacidade, &cliente.Cnpj)
+			sec.CheckInternalServerError(err, w)
+			cliente.Order = i
+			i++
+			clientes = append(clientes, cliente)
+		}
+		var page mdl.PageClientes
+		page.AppName = mdl.AppName
+		page.Clientes = clientes
+		page.Title = "Clientes"
+		page.LoggedUser = BuildLoggedUser(GetUserInCookie(w, r))
+		var tmpl = template.Must(template.ParseGlob("tiles/clientes/*"))
+		tmpl.ParseGlob("tiles/*")
+		tmpl.ExecuteTemplate(w, "Main-Clientes", page)
+		sec.CheckInternalServerError(err, w)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	var page mdl.PageClientes
-	page.AppName = mdl.AppName
-	page.Clientes = clientes
-	page.Title = "Clientes"
-	page.LoggedUser = BuildLoggedUser(GetUserInCookie(w, r))
-	var tmpl = template.Must(template.ParseGlob("tiles/clientes/*"))
-	tmpl.ParseGlob("tiles/*")
-	tmpl.ExecuteTemplate(w, "Main-Clientes", page)
-	sec.CheckInternalServerError(err, w)
 }
-
