+++ handlers/handlersFeatures.go
@@ -11,9 +11,8 @@ import (
 )
 
 func CreateFeatureHandler(w http.ResponseWriter, r *http.Request) {
-	sec.IsAuthenticated(w, r)
-	log.Println("Create Feature")
-	if r.Method == "POST" {
+	if r.Method == "POST" && sec.IsAuthenticated(w, r) {
+		log.Println("Create Feature")
 		name := r.FormValue("Name")
 		code := r.FormValue("Code")
 		sqlStatement := "INSERT INTO features(name, code) VALUES ($1, $2) RETURNING id"
@@ -25,14 +24,15 @@ func CreateFeatureHandler(w http.ResponseWriter, r *http.Request) {
 		}
 		sec.CheckInternalServerError(err, w)
 		log.Println("INSERT: Id: " + strconv.Itoa(id) + " | Name: " + name + " | Code: " + code)
+		http.Redirect(w, r, route.FeaturesRoute, 301)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	http.Redirect(w, r, route.FeaturesRoute, 301)
 }
 
 func UpdateFeatureHandler(w http.ResponseWriter, r *http.Request) {
-	sec.IsAuthenticated(w, r)
-	log.Println("Update Feature")
-	if r.Method == "POST" {
+	if r.Method == "POST" && sec.IsAuthenticated(w, r) {
+		log.Println("Update Feature")
 		id := r.FormValue("Id")
 		name := r.FormValue("Name")
 		code := r.FormValue("Code")
@@ -45,14 +45,15 @@ func UpdateFeatureHandler(w http.ResponseWriter, r *http.Request) {
 		sec.CheckInternalServerError(err, w)
 		updtForm.Exec(name, code, id)
 		log.Println("UPDATE: Id: " + id + " | Name: " + name + " | Code: " + code)
+		http.Redirect(w, r, route.FeaturesRoute, 301)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	http.Redirect(w, r, route.FeaturesRoute, 301)
 }
 
 func DeleteFeatureHandler(w http.ResponseWriter, r *http.Request) {
-	sec.IsAuthenticated(w, r)
-	log.Println("Delete Feature")
-	if r.Method == "POST" {
+	if r.Method == "POST" && sec.IsAuthenticated(w, r) {
+		log.Println("Delete Feature")
 		id := r.FormValue("Id")
 		sqlStatement := "DELETE FROM features WHERE id=$1"
 		deleteForm, err := Db.Prepare(sqlStatement)
@@ -62,8 +63,10 @@ func DeleteFeatureHandler(w http.ResponseWriter, r *http.Request) {
 		deleteForm.Exec(id)
 		sec.CheckInternalServerError(err, w)
 		log.Println("DELETE: Id: " + id)
+		http.Redirect(w, r, route.FeaturesRoute, 301)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	http.Redirect(w, r, route.FeaturesRoute, 301)
 }
 
 func DeleteFeaturesByRoleHandler(roleId string) {
@@ -89,28 +92,31 @@ func DeleteFeaturesHandler(diffDB []mdl.Feature) {
 
 func ListFeaturesHandler(w http.ResponseWriter, r *http.Request) {
 	log.Println("List Features")
-	sec.IsAuthenticated(w, r)
-	rows, err := Db.Query("SELECT id, name, code FROM features order by id asc")
-	sec.CheckInternalServerError(err, w)
-	var features []mdl.Feature
-	var feature mdl.Feature
-	var i = 1
-	for rows.Next() {
-		err = rows.Scan(&feature.Id, &feature.Name, &feature.Code)
+	if sec.IsAuthenticated(w, r) {
+		rows, err := Db.Query("SELECT id, name, code FROM features order by id asc")
 		sec.CheckInternalServerError(err, w)
-		feature.Order = i
-		i++
-		features = append(features, feature)
+		var features []mdl.Feature
+		var feature mdl.Feature
+		var i = 1
+		for rows.Next() {
+			err = rows.Scan(&feature.Id, &feature.Name, &feature.Code)
+			sec.CheckInternalServerError(err, w)
+			feature.Order = i
+			i++
+			features = append(features, feature)
+		}
+		var page mdl.PageFeatures
+		page.AppName = mdl.AppName
+		page.Features = features
+		page.Title = "Funcionalidades"
+		page.LoggedUser = BuildLoggedUser(GetUserInCookie(w, r))
+		var tmpl = template.Must(template.ParseGlob("tiles/features/*"))
+		tmpl.ParseGlob("tiles/*")
+		tmpl.ExecuteTemplate(w, "Main-Features", page)
+		sec.CheckInternalServerError(err, w)
+	} else {
+		http.Redirect(w, r, "/logout", 301)
 	}
-	var page mdl.PageFeatures
-	page.AppName = mdl.AppName
-	page.Features = features
-	page.Title = "Funcionalidades"
-	page.LoggedUser = BuildLoggedUser(GetUserInCookie(w, r))
-	var tmpl = template.Must(template.ParseGlob("tiles/features/*"))
-	tmpl.ParseGlob("tiles/*")
-	tmpl.ExecuteTemplate(w, "Main-Features", page)
-	sec.CheckInternalServerError(err, w)
 }
 
 // AJAX
